fix: extract isomorphic address-labels-shared, sanitize Gnosis Safe import path
