+++ ui-dashboard/src/components/address-label-editor.tsx
@@ -2,7 +2,7 @@
 
 import { useRef, useEffect, useState, useMemo } from "react";
 import { useAddressLabels } from "@/components/address-labels-provider";
-import type { AddressEntry } from "@/lib/address-labels";
+import type { AddressEntry } from "@/lib/address-labels-shared";
 import { TagInput } from "@/components/tag-input";
 import { SUGGESTED_TAGS, getUsedTags } from "@/lib/tag-suggestions";
